pyproject.toml
[tool.black]
line-length = 79
target-version = ["py38"]

[tool.isort]
known_third_party = ["aiofiles", "aioredis", "dotenv", "invoke", "pytest", "starlette"]
profile = "black"
src_paths = ["line_server"]

[tool.poetry]
name = "line-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
starlette = "^0.20.3"
uvicorn = "^0.17.6"
aioredis = "^2.0.1"
asyncio = "^3.4.3"
python-dotenv = "^0.20.0"
aiofiles = "^0.8.0"
requests = "^2.28.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
invoke = "^1.7.1"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.3"
types-redis = "^4.2.8"
types-aiofiles = "^0.8.8"
pre-commit = "^2.19.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
